Replaced MyTripViewController with MyTripTableViewController and changed out tab bar segues
