--- travely/MyTripsViewController.swift
@@ -1,35 +1,0 @@
-//
-//  MyTripsViewController.swift
-//  travely
-//
-//  Created by Alexandra Leonidova on 3/11/18.
-//  Copyright © 2018 University of San Diego. All rights reserved.
-//
-
-import UIKit
-
-class MyTripsViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-        // Do any additional setup after loading the view.
-    }
-
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
-    
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destinationViewController.
-        // Pass the selected object to the new view controller.
-    }
-    */
-
-}
